requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
linkedin-api>=2.0.0
selenium>=4.15.0
fastapi>=0.104.0
pydantic>=2.4.0
python-jose>=3.3.0
pytest>=7.4.0
